+++ src/esm/index.js
@@ -96,11 +96,7 @@ export { Grid, GridCol } from './components/grid/grid';
 export { Row } from './components/grid/row';
 export { SideGrid } from './components/grid/side-grid';
 export { StickyContainer } from './components/grid/sticky-container';
-export { AdaptableGrid } from './components/grid/adaptable-grid'; // Headers
-
-export { HeaderLogo } from './components/headers/header-logo';
-export { HeaderTitles } from './components/headers/header-titles';
-export { Header, HeaderItems, HeaderItem } from './components/headers/header'; // Heroes
+export { AdaptableGrid } from './components/grid/adaptable-grid'; // Heroes
 
 export { Hero } from './components/heroes/hero'; // Images